+++ modules/fundamental/tests/fundamental.rs
@@ -1,0 +1,1 @@
+mod sbom;
